1.1

+++ src/Modules/Gangs.cs
@@ -6,10 +6,12 @@ using DEA.Database.Repository;
 using System.Linq;
 using DEA.Database.Models;
 using MongoDB.Driver;
+using DEA.Resources;
+using DEA.Services;
 
 namespace DEA.Modules
 {
-    public class Gangs : ModuleBase<SocketCommandContext>
+    public class Gangs : DEAModule
     {
 
         [Command("CreateGang")]
@@ -23,7 +25,7 @@ namespace DEA.Modules
                 throw new Exception($"You do not have {Config.GANG_CREATION_COST.ToString("C", Config.CI)}. Balance: {user.Cash.ToString("C", Config.CI)}.");
             var gang = GangRepository.CreateGang(Context.User.Id, Context.Guild.Id, name);
             await UserRepository.EditCashAsync(Context, -Config.GANG_CREATION_COST);
-            await ReplyAsync($"{Context.User.Mention}, You have successfully created the {gang.Name} gang!");
+            await Reply($"You have successfully created the {gang.Name} gang!");
         }
 
         [Command("AddGangMember")]
@@ -35,40 +37,18 @@ namespace DEA.Modules
             if (GangRepository.InGang(user.Id, Context.Guild.Id)) throw new Exception("This user is already in a gang.");
             if (GangRepository.IsFull(Context.User.Id, Context.Guild.Id)) throw new Exception("Your gang is already full!");
             GangRepository.AddMember(Context.User.Id, Context.Guild.Id, user.Id);
-            await ReplyAsync($"{user} is now a new member of your gang!");
+            await Reply($"{user} is now a new member of your gang!");
             var channel = await user.CreateDMChannelAsync();
-            await channel.SendMessageAsync($"Congrats! You are now a member of {GangRepository.FetchGang(Context).Name}!");
+            await ResponseMethods.DM(channel, $"Congrats! You are now a member of {GangRepository.FetchGang(Context).Name}!");
         }
-
-        /*[Command("JoinGang", RunMode = RunMode.Async)]
-        [Require("nogang")]
-        [Summary("Allows you to request to join a gang.")]
-        [Remarks("JoinGang <@GangMember>")]
-        private async Task JoinGang(IGuildUser user)
-        {
-            if (!GangRepository.InGang(user.Id, Context.Guild.Id)) throw new Exception("This user is not in a gang.");
-            if (GangRepository.IsFull(user.Id, Context.Guild.Id)) throw new Exception("This gang is already full!");
-            var gang = GangRepository.FetchGang(user.Id, user.GuildId);
-            var channel = await user.CreateDMChannelAsync();
-            await channel.SendMessageAsync($"{Context.User} has requested to join your gang. Reply with \"agree\" within the next 30 seconds to accept this request.");
-            await ReplyAsync($"{Context.User.Mention}, The leader of {gang.Name} has been successfully informed of your request to join.");
-            var response = await WaitForMessage(Context.User, Context.Channel, TimeSpan.FromSeconds(30));
-            await ReplyAsync($"{response.Content}");
-            if (response.Content.ToLower() == "agree")
-            {
-                GangRepository.AddMember(gang.LeaderId, user.GuildId, Context.User.Id);
-                await channel.SendMessageAsync($"{Context.User} is now a new member of your gang!");
-                var informingChannel = await Context.User.CreateDMChannelAsync();
-                await informingChannel.SendMessageAsync($"{Context.User.Mention}, Congrats! You are now a member of {gang.Name}!");
-            }
-        }*/
 
         [Command("Gang")]
         [Summary("Gives you all the info about any gang.")]
         [Remarks("Gang [Gang name]")]
         public async Task Gang([Remainder] string gangName = null)
         {
-            if (gangName == null && !GangRepository.InGang(Context.User.Id, Context.Guild.Id)) throw new Exception($"You are not in a gang.");
+            if (gangName == null && !GangRepository.InGang(Context.User.Id, Context.Guild.Id))
+                throw new Exception($"You are not in a gang.");
             Gang gang;
             if (gangName == null) gang = GangRepository.FetchGang(Context);
             else gang = GangRepository.FetchGang(gangName, Context.Guild.Id);
@@ -77,19 +57,10 @@ namespace DEA.Modules
             if (Context.Guild.GetUser(gang.LeaderId) != null) leader = $"<@{gang.LeaderId}>";
             foreach (var member in gang.Members)
                 if (Context.Guild.GetUser(member) != null) members += $"<@{member}>, ";
-            var InterestRate = 0.025f + ((gang.Wealth / 100) * .000075f);
-            if (InterestRate > 0.1) InterestRate = 0.1f;
             if (members.Length > 2) members = $"__**Members:**__ {members.Substring(0, members.Length - 2)}\n";
-            var builder = new EmbedBuilder()
-            {
-                Title = gang.Name,
-                Color = new Color(0x00AE86),
-                Description = $"__**Leader:**__ {leader}\n" +
-                              members +
-                              $"__**Wealth:**__ {gang.Wealth.ToString("C", Config.CI)}\n" +
-                              $"__**Interest rate:**__ {InterestRate.ToString("P")}"
-            };
-            await ReplyAsync("", embed: builder);
+            var description = $"__**Leader:**__ {leader}\n" + members + $"__**Wealth:**__ {gang.Wealth.ToString("C", Config.CI)}\n" +
+                              $"__**Interest rate:**__ {Services.Math.CalculateIntrestRate(gang.Wealth).ToString("P")}";
+            await Send(description, gang.Name);
         }
 
         [Command("GangLb")]
@@ -98,24 +69,20 @@ namespace DEA.Modules
         [Remarks("Gangs")]
         public async Task Ganglb()
         {
-            var gangs = await DEABot.Gangs.FindAsync(y => y.GuildId == Context.Guild.Id);
-            var sortedGangs = (await gangs.ToListAsync()).OrderByDescending(x => x.Wealth).ToList();
-            string message = "```asciidoc\n= The Wealthiest Gangs =\n";
-            int longest = 0;
+            var gangs = DEABot.Gangs.Find(y => y.GuildId == Context.Guild.Id).ToList();
+
+            if (gangs.Count == 0) throw new Exception("There aren't any gangs yet.");
+
+            var sortedGangs = gangs.OrderByDescending(x => x.Wealth).ToList();
+            string description = "";
 
             for (int i = 0; i < sortedGangs.Count(); i++)
             {
                 if (i + 1 >= Config.GANGSLB_CAP) break;
-                if (sortedGangs[i].Name.Length > longest) longest = $"{i + 1}. {sortedGangs[i].Name}".Length;
+                description += $"{i + 1}. {sortedGangs[i].Name}: {sortedGangs[i].Wealth.ToString("C", Config.CI)}\n";
             }
 
-            for (int i = 0; i < sortedGangs.Count(); i++)
-            {
-                if (i + 1 >= Config.GANGSLB_CAP) break;
-                message += $"{i + 1}. {sortedGangs[i].Name}".PadRight(longest + 2) + $" :: {sortedGangs[i].Wealth.ToString("C", Config.CI)}\n";
-            }
-
-            await ReplyAsync($"{message}```");
+            await Send(description, "The Wealthiest Gangs");
         }
 
         [Command("LeaveGang")]
@@ -130,9 +97,9 @@ namespace DEA.Modules
                 throw new Exception($"You may not leave a gang if you are the owner. Either destroy the gang with the `{prefix}DestroyGang` command, or " +
                                     $"transfer the ownership of the gang to another member with the `{prefix}TransferLeadership` command.");
             GangRepository.RemoveMember(Context.User.Id, Context.Guild.Id);
-            await ReplyAsync($"{Context.User.Mention}, You have successfully left {gang.Name}");
+            await Reply($"You have successfully left {gang.Name}");
             var channel = await Context.Client.GetUser(gang.LeaderId).CreateDMChannelAsync();
-            await channel.SendMessageAsync($"{Context.User} has left {gang.Name}.");
+            await ResponseMethods.DM(channel, $"{Context.User} has left {gang.Name}.");
         }
 
         [Command("KickGangMember")]
@@ -141,13 +108,15 @@ namespace DEA.Modules
         [Remarks("KickGangMember")]
         public async Task KickFromGang([Remainder] IGuildUser user)
         {
+            if (user.Id == Context.User.Id)
+                throw new Exception("You may not kick yourself!");
             if (!GangRepository.IsMemberOf(Context.User.Id, Context.Guild.Id, user.Id))
                 throw new Exception("This user is not a member of your gang!");
             var gang = GangRepository.FetchGang(Context);
             GangRepository.RemoveMember(user.Id, Context.Guild.Id);
-            await ReplyAsync($"{Context.User.Mention}, You have successfully kicked {user} from {gang.Name}");
+            await Reply($"You have successfully kicked {user} from {gang.Name}");
             var channel = await user.CreateDMChannelAsync();
-            await channel.SendMessageAsync($"You have been kicked from {gang.Name}.");
+            await ResponseMethods.DM(channel, $"You have been kicked from {gang.Name}.");
         }
 
         [Command("DestroyGang")]
@@ -157,7 +126,7 @@ namespace DEA.Modules
         public async Task DestroyGang()
         {
             GangRepository.DestroyGang(Context.User.Id, Context.Guild.Id);
-            await ReplyAsync($"{Context.User.Mention}, You have successfully destroyed your gang.");
+            await Reply($"You have successfully destroyed your gang.");
         }
 
         [Command("ChangeGangName")]
@@ -170,11 +139,11 @@ namespace DEA.Modules
             var user = UserRepository.FetchUser(Context);
             if (user.Cash < Config.GANG_NAME_CHANGE_COST)
                 throw new Exception($"You do not have {Config.GANG_NAME_CHANGE_COST.ToString("C", Config.CI)}. Balance: {user.Cash.ToString("C", Config.CI)}.");
-            var gangs = await (await DEABot.Gangs.FindAsync(y => y.GuildId == Context.Guild.Id)).ToListAsync();
-            if (!gangs.Any(x => x.Name.ToLower() == name.ToLower())) throw new Exception($"There is already a gang by the name {name}.");
+            var gangs = DEABot.Gangs.Find(y => y.GuildId == Context.Guild.Id).ToList();
+            if (gangs.Any(x => x.Name.ToLower() == name.ToLower())) throw new Exception($"There is already a gang by the name {name}.");
             await UserRepository.EditCashAsync(Context, -Config.GANG_NAME_CHANGE_COST);
             GangRepository.Modify(DEABot.GangUpdateBuilder.Set(x => x.Name, name), Context);
-            await ReplyAsync($"You have successfully changed your gang name to {name} at the cost of {Config.GANG_NAME_CHANGE_COST.ToString("C", Config.CI)}.");
+            await Reply($"You have successfully changed your gang name to {name} at the cost of {Config.GANG_NAME_CHANGE_COST.ToString("C", Config.CI)}.");
         }
 
         [Command("TransferLeadership")]
@@ -195,14 +164,14 @@ namespace DEA.Modules
                         DEABot.GangUpdateBuilder.Set(x => x.Members, gang.Members)), Context);
                     break;
                 }
-            await ReplyAsync($"{Context.User.Mention}, You have successfully transferred the leadership of {gang.Name} to {user.Mention}");
+            await Reply($"You have successfully transferred the leadership of {gang.Name} to {user.Mention}");
         }
 
         [Command("Deposit")]
         [Require(Attributes.InGang)]
         [Summary("Deposit cash into your gang's funds.")]
         [Remarks("Deposit <Cash>")]
-        public async Task Deposit(double cash)
+        public async Task Deposit(decimal cash)
         {
             var user = UserRepository.FetchUser(Context);
             if (cash < Config.MIN_DEPOSIT) throw new Exception($"The lowest deposit is {Config.MIN_DEPOSIT.ToString("C", Config.CI)}.");
@@ -210,8 +179,8 @@ namespace DEA.Modules
             await UserRepository.EditCashAsync(Context, -cash);
             var gang = GangRepository.FetchGang(Context);
             GangRepository.Modify(DEABot.GangUpdateBuilder.Set(x => x.Wealth, gang.Wealth + cash), Context.User.Id, Context.Guild.Id);
-            await ReplyAsync($"{Context.User.Mention}, You have successfully deposited {cash.ToString("C", Config.CI)}. " +
-                             $"{gang.Name}'s Wealth: {(gang.Wealth + cash).ToString("C", Config.CI)}");
+            await Reply($"You have successfully deposited {cash.ToString("C", Config.CI)}. " +
+                        $"{gang.Name}'s Wealth: {(gang.Wealth + cash).ToString("C", Config.CI)}");
         }
 
         [Command("Withdraw")]
@@ -219,7 +188,7 @@ namespace DEA.Modules
         [RequireCooldown]
         [Summary("Withdraw cash from your gang's funds.")]
         [Remarks("Withdraw <Cash>")]
-        public async Task Withdraw(double cash)
+        public async Task Withdraw(decimal cash)
         {
             var gang = GangRepository.FetchGang(Context);
             var user = UserRepository.FetchUser(Context);
@@ -230,8 +199,8 @@ namespace DEA.Modules
             UserRepository.Modify(DEABot.UserUpdateBuilder.Set(x => x.Withdraw, DateTime.UtcNow), Context);
             GangRepository.Modify(DEABot.GangUpdateBuilder.Set(x => x.Wealth, gang.Wealth - cash), Context.User.Id, Context.Guild.Id);
             await UserRepository.EditCashAsync(Context, +cash);
-            await ReplyAsync($"{Context.User.Mention}, You have successfully withdrawn {cash.ToString("C", Config.CI)}. " +
-                             $"{gang.Name}'s Wealth: {(gang.Wealth - cash).ToString("C", Config.CI)}");
+            await Reply($"You have successfully withdrawn {cash.ToString("C", Config.CI)}. " +
+                        $"{gang.Name}'s Wealth: {(gang.Wealth - cash).ToString("C", Config.CI)}");
         }
 
     }
